fix: User clas constructor Syntax error

+++ model/Level.java
@@ -3,13 +3,13 @@ package model;
 enum Level {
     EMPLOYEE(1), CLIENT(2);
 
-    private final byte level;
+    private final int level;
 
-    Level (byte level) {
+    Level (int level) {
         this.level = level;
     }
 
-	public byte get() {
+	public int get() {
 		return (this.level);
 	}
 }
